+++ pcap4j-core/src/main/java/org/pcap4j/core/PcapNativeException.java
@@ -15,6 +15,8 @@ public final class PcapNativeException extends Exception {
 
   private static final long serialVersionUID = -6458526492950674556L;
 
+  private Integer returnCode = null;
+
   /**
    *
    */
@@ -33,6 +35,16 @@ public final class PcapNativeException extends Exception {
   /**
    *
    * @param message
+   * @param returnCode
+   */
+  public PcapNativeException(String message, Integer returnCode){
+    super(message);
+    this.returnCode = returnCode;
+  }
+
+  /**
+   *
+   * @param message
    * @param cause
    */
   public PcapNativeException(String message, Throwable cause){
@@ -45,6 +57,19 @@ public final class PcapNativeException extends Exception {
    */
   public PcapNativeException(Throwable cause){
       super(cause);
+  }
+
+  /**
+   *
+   * @return the return code if the native function which caused
+   *         this exception returned it; otherwise null.
+   */
+  public Integer getReturnCode() {
+    return returnCode;
+  }
+
+  void setReturnCode(Integer returnCode) {
+    this.returnCode = returnCode;
   }
 
 }